Create Leetcode DSA : Next Permutation
 LeetCode

Today, I want to share a fascinating algorithm called the "Next Permutation," a powerful tool in the world of computer science and programming. This algorithm is particularly useful when dealing with arrangements of elements, known as permutations.
The Next Permutation Algorithm

Permutations represent the different ways we can arrange a set of elements. The "Next Permutation" algorithm allows us to find the lexicographically next arrangement of elements in a sequence. It's like turning the page to the next word in a dictionary.

Here's a high-level overview of the algorithm:

    Identify the Break Point: Starting from the right, find the first pair of elements where the left element is smaller than the right element. This is our "break point."
    Handle Special Cases: If no break point is found, the sequence is already at its maximum lexicographical order. In this case, we simply reverse the entire sequence.
    Swap and Reverse: If there is a break point, we find the smallest element to the right of the break point that is larger than the element at the break point. Swap these two elements and then reverse the subsequence to the right of the break point.